Type Alias dotfiles_actions::create::action::FakeCreateAction

pub type FakeCreateAction<'a> = CreateAction<'a, FakeFileSystem>;

A Fake create action that works on a fake test filesystem.

Aliased Type

struct FakeCreateAction<'a> { /* private fields */ }

Implementations

impl<'a, F: FileSystem> CreateAction<'a, F>

pub fn directory(&self) -> &String

Directory to create. Can be absolute or relative.

pub fn create_parent_dirs(&self) -> &bool

Force creation of the directory and all its parents if they do not exist already.

Setting create_parent_dirs to true is equivalent to using the -p flag in mkdir.

pub fn current_dir(&self) -> &PathBuf

Current directory that will be used to determine relative file locations if necessary. It must match the parent directory of the configuration file that declared this action.

impl<'a, F: FileSystem> CreateAction<'a, F>

pub fn new( fs: &'a F, skip_in_ci: bool, directory: String, create_parent_dirs: bool, current_dir: PathBuf ) -> Result<Self, DotfilesError>

Constructs a new instance of CreateAction