dotfiles_actions::create

Module directive

This module defines CreateDirective.

Structs

CreateDirective

A directive that can build CreateActions to create directories in the filesystem.

Constants

CREATE_PARENT_DIRS_SETTING

Constant for the name of the create_parent_dirs Setting which forces to create all parent directories if necessary.

DIRECTIVE_NAME

Constant for the name of the create directive.

DIR_SETTING

Constant for the name of the directory argument that contains the name of the directory to create

Functions

init_directive_data

Initializes the default configuration for the CreateDirective

Type Aliases

FakeCreateDirective

CreateDirective that uses the native [FakeFileSystem] for testing.

NativeCreateDirective

CreateDirective that uses the native [OsFileSystem].