dotfiles_actions::install_command

Trait InstallCommand

pub trait InstallCommand<F: Display> {
    // Required methods
    fn base_command(&self) -> Exec;
    fn args(&self) -> &Vec<String>;
    fn action_description(&self) -> &str;
    fn action_name(&self) -> &str;
    fn items(&self) -> &Vec<F>;

    // Provided methods
    fn formatted_item_list(&self) -> Vec<String> { ... }
    fn execute(&self) -> Result<(), DotfilesError> { ... }
}

Trait that represents a command that installs an item using an action (brew install, brew install cask, apt install…)

Required Methods

fn base_command(&self) -> Exec

The base command to run

fn args(&self) -> &Vec<String>

The arguments to pass to the command

fn action_description(&self) -> &str

The description of the action to run, i.e. “Installing cask”

fn action_name(&self) -> &str

The name of the action to run, i.e. “cask”

fn items(&self) -> &Vec<F>

The item actually being installed, for example a homebrew formula.

Provided Methods

fn formatted_item_list(&self) -> Vec<String>

a list of items to display

fn execute(&self) -> Result<(), DotfilesError>

Runs the command to execut