Module dotfiles_actions::link::action

This module contains the LinkAction that creates a new symlink when executed

Structs

LinkAction

LinkAction creates a new symlink path that points to target.

Type Aliases

FakeLinkAction

A Fake create action that works on a fake test filesystem.

NativeLinkAction

A native create action that works on the real filesystem.