Type Alias dotfiles_actions::link::action::NativeLinkAction

pub type NativeLinkAction<'a> = LinkAction<'a, OsFileSystem>;

A native create action that works on the real filesystem.

Aliased Type

struct NativeLinkAction<'a> { /* private fields */ }

Implementations

impl<'a, F: FileSystem + UnixFileSystem> LinkAction<'a, F>

pub fn path(&self) -> &String

Path of the new symlink

pub fn target(&self) -> &String

Path that the symlink points to.

impl<'a, F: FileSystem + UnixFileSystem> LinkAction<'a, F>

pub fn relink(&self) -> bool

Force to re-create the symlink if it exists already

pub fn force(&self) -> bool

Force to replace an existing file or directory when executed.

pub fn create_parent_dirs(&self) -> bool

Create all parent directories if they do not exist already

pub fn ignore_missing_target(&self) -> bool

Succeed even if target doesn’t point to an existing file or dir.

pub fn resolve_symlink_target(&self) -> bool

If the target is another symlink, resolve the ultimate concrete file or directory that it points to and make it the target

impl<'a, F: FileSystem + UnixFileSystem> LinkAction<'a, F>

pub fn new( fs: &'a F, path: String, target: String, context_settings: &Settings, defaults: &Settings, current_dir: PathBuf ) -> Result<Self, DotfilesError>

Constructs a new LinkAction