Crate dotfiles_actions

This crate contains all concrete actions and directives. They are contained in each of the modules in this crate:

brew: A set of actions and directive that act as wrappers around homebrew.

create: Creates a new directory

exec: runs a command in the shell

homebrew_install: installs homebrew.

link: Creates a symlink to a file or directory

ohmyzsh_install: installs ohmyzsh

Modules

apt

This module contains the AptAction and AptDirective

brew

This module contains the BrewAction and BrewDirective

create

This module contains the CreateAction and CreateDirective classes that are used by the runtime to create a new directory.

exec

This module contains the ExecAction and ExecDirective used to execute a command on the shell.

filesystem

Module that contains interfaces common to directives that use filesystems.

homebrew_install

This module contains the HomebrewInstallAction

install_command

Module that contains common code for all commands that install a package

link

This module contains the LinkAction and LinkDirective classes that are used by the runtime to create a symlink.

ohmyzsh_install

This module contains the OhMyZshInstallAction