Module dotfiles_actions::brew::directive

This module defines BrewDirective.

Structs

BrewDirective

A directive that can build BrewActions to install formulae, casks

Constants

CASK_SETTING

The string that identifies the list of casks to install

DIRECTIVE_NAME

Name of the Brew directive

FORCE_CASKS_SETTING

force casks to deal with previously installed apps

FORMULA_SETTING

The string that identifies the list of formulae to install

TAP_SETTING

The string that identifies the list of taps to install

Functions

init_directive_data

Initialize the defaults for the BrewDirective.